requirements.txt
dvc==3.25.0
evidently==0.4.5
fastparquet==2023.2.0
gitpython==3.1.40
jupyterlab
mlflow==2.7.1
numpy==1.26.0
pendulum
pandas==2.1.1
pyarrow==13.0.0
requests==2.31.0
scikit-learn==1.3.1
psycopg2-binary==2.9.9
python-gitlab==3.15.0
sqlalchemy
tqdm
